+++ src/renderer/src/bootstrap.test.ts
@@ -37,6 +37,7 @@ function noopDeps(): CommandDeps {
     jumpToNextAttention: () => {},
     openAbout: () => {},
     checkForUpdates: () => {},
+    openFile: () => {},
   };
 }